+++ R/pca_rob.R
@@ -1,4 +1,4 @@
-pcarob <- function(X, ncomp, nsim = 1500, alpha = .30, step2 = TRUE, 
+pca_rob <- function(X, ncomp, nsim = 1500, alpha = .30, step2 = TRUE, 
                                         weights = NULL, ...) {
     
     X <- .matrix(X)