+++ R/pls_rob.R
@@ -15,8 +15,8 @@ pls_rob <- function(X, Y, ncomp, ncompw = 10, alpha = .30,
     
     ncompw <- min(ncompw, n, p)
     
-    fm <- pcarob(X, ncompw, typcut = typcut, weights = weights, ...)
-    #fm <- pca.sph(X, min(ncompw, n, p), ...)
+    fm <- pca_rob(X, ncompw, typcut = typcut, weights = weights, ...)
+    #fm <- pca_sph(X, min(ncompw, n, p), ...)
     r <- outsdod(fm, X, typcut = typcut)
     wx <- .talworth(r, quantile(r, 1 - alpha))